Preparing Model for Deployment
