+++ .gitignore
@@ -1,2 +1,3 @@
 /stck
 *.csv
+*.pkl
